pyproject.toml
[project]
name = "5-rag-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.0",
    "chardet>=5.2.0",
    "cryptography>=46.0.3",
    "docx2txt>=0.9",
    "filetype>=1.2.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "jieba>=0.42.1",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-huggingface>=1.2.0",
    "langchain-milvus>=0.3.2",
    "langchain-openai>=1.1.6",
    "minio>=7.2.20",
    "numpy>=2.4.0",
    "pdfplumber>=0.11.9",
    "pymysql>=1.1.2",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sqlalchemy>=2.0.45",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
